refactor: минорные исправления
