--- src/main/java/ru/elessarov/workout_notebook_bot/utils/AdminProperties.java
@@ -1,17 +1,0 @@
-package ru.elessarov.workout_notebook_bot.utils;
-
-import lombok.Getter;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
-
-@Component
-@Getter
-public class AdminProperties {
-
-    public static String ADMIN_ID;
-
-    @Value("${constants.admin_id}")
-    public void setAdminStatic(String adminId) {
-        AdminProperties.ADMIN_ID = adminId;
-    }
-}
