+++ src/main/java/ru/elessarov/workout_notebook_bot/utils/BotUtils.java
@@ -13,10 +13,10 @@ import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
 import java.util.List;
 
-import static ru.elessarov.workout_notebook_bot.utils.AdminProperties.ADMIN_ID;
 
 @UtilityClass
 public class BotUtils {
+
     private static final DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm:ss");
     private static final DateTimeFormatter dateFormatter = DateTimeFormatter.ofPattern("yyyy-MM-dd");
 
@@ -44,14 +44,14 @@ public class BotUtils {
         sendMessage.setReplyMarkup(inlineKeyboardMarkup);
     }
 
-    public static SendMessage notifyAdmin(final UserEntity user, final int months) {
-        var message = new SendMessage(ADMIN_ID, "@%s want to sub for %d months".formatted(user.getUsername(), months));
+    public static SendMessage notifyAdmin(final UserEntity user, final int months, final String adminId) {
+        var message = new SendMessage(adminId, "@%s want to sub for %d months".formatted(user.getUsername(), months));
         addKeyboard(message, List.of("Подтвердить оплату, %d".formatted(user.getSubscribe().getId())));
         return message;
     }
 
-    public static boolean isUserAdmin(final User user) {
-        return String.valueOf(user.getId()).equals(ADMIN_ID);
+    public static boolean isUserAdmin(final User user, final String adminId) {
+        return String.valueOf(user.getId()).equals(adminId);
     }
 
     public static int extractSubId(final String value) {
