--- src/main/resources/db/changelog/data/2024_03_09_insert_exercise_examples.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<databaseChangeLog xmlns="http://www.liquibase.org/xml/ns/dbchangelog"
-                   xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-                   xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog
-                   https://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-latest.xsd">
-
-    <changeSet id="20240309-01" author="Fisanov_E">
-        <loadData file="exercises_list.csv" tableName="exercises"/>
-    </changeSet>
-</databaseChangeLog>
